+++ lib/hexapdf/type/acro_form/appearance_generator.rb
@@ -90,6 +90,10 @@ module HexaPDF
 
         # Creates the appropriate appearances for check boxes.
         #
+        # The unchecked box is always represented by the appearance with the key /Off. If there is
+        # more than one other key besides the /Off key, the first one is used for the appearance of
+        # the checked box.
+        #
         # For unchecked boxes an empty rectangle is drawn. When checked, a symbol from the
         # ZapfDingbats font is placed inside the rectangle. How this is exactly done depends on the
         # following values:
@@ -120,7 +124,9 @@ module HexaPDF
         #   widget.marker_style(style: :cross)
         #   # => no visible rectangle, gray background, cross mark when checked
         def create_check_box_appearances
-          unless @widget.appearance_dict&.normal_appearance&.value&.size == 2
+          appearance_keys = @widget.appearance_dict&.normal_appearance&.value&.keys || []
+          on_name = (appearance_keys - [:Off]).first
+          unless on_name
             raise HexaPDF::Error, "Widget of check box doesn't define name for on state"
           end
           border_style = @widget.border_style
@@ -132,7 +138,7 @@ module HexaPDF
             @document.add({Type: :XObject, Subtype: :Form, BBox: [0, 0, rect.width, rect.height]})
           apply_background_and_border(border_style, off_form.canvas)
 
-          on_form = @widget.appearance_dict.normal_appearance[@field.check_box_on_name] =
+          on_form = @widget.appearance_dict.normal_appearance[on_name] =
             @document.add({Type: :XObject, Subtype: :Form, BBox: [0, 0, rect.width, rect.height]})
           canvas = on_form.canvas
           apply_background_and_border(border_style, canvas)
@@ -169,11 +175,12 @@ module HexaPDF
         #   widget.marker_style(style: :circle, size: 0, color: 0)
         #   # => default appearance
         def create_radio_button_appearances
-          unless @widget.appearance_dict&.normal_appearance&.value&.size == 2
-            raise HexaPDF::Error, "Widget of radio button doesn't define unique name for on state"
+          appearance_keys = @widget.appearance_dict&.normal_appearance&.value&.keys || []
+          on_name = (appearance_keys - [:Off]).first
+          unless on_name
+            raise HexaPDF::Error, "Widget of radio button doesn't define name for on state"
           end
 
-          on_name = (@widget.appearance_dict.normal_appearance.value.keys - [:Off]).first
           border_style = @widget.border_style
           marker_style = @widget.marker_style